+++ 2DPhysicEngine/FabricSimulation/Graphics.cpp
@@ -132,7 +132,7 @@ void Graphics::DrawFillRect(int x, int y, int width, int height, Uint32 color)
  */
 void Graphics::DrawPolygon(int x, int y, const std::vector<Vector2D>& vertices, Uint32 color)
 {
-    for (int i = 0; i < vertices.size(); i++)
+    for (int i = 0; i < static_cast<int>(vertices.size()); i++)
     {
         const int currIndex = i;
         const int nextIndex = (i + 1) % vertices.size();
@@ -150,7 +150,7 @@ void Graphics::DrawFillPolygon(int x, int y, const std::vector<Vector2D>& vertic
     std::vector<short> vx;
     std::vector<short> vy;
 
-    for (int i = 0; i < vertices.size(); i++)
+    for (int i = 0; i < static_cast<int>(vertices.size()); i++)
     {
         vx.push_back(static_cast<int>(vertices[i].x));
     }
