+++ 2DPhysicEngine/Ragdoll/Graphics.cpp
@@ -15,7 +15,8 @@ int Graphics::Height() {
 }
 
 bool Graphics::OpenWindow() {
-    if (SDL_Init(SDL_INIT_EVERYTHING) != 0) {
+    if (SDL_Init(SDL_INIT_EVERYTHING) != 0) 
+    {
         std::cerr << "Error initializing SDL" << std::endl;
         return false;
     }
@@ -24,41 +25,49 @@ bool Graphics::OpenWindow() {
     windowWidth = display_mode.w;
     windowHeight = display_mode.h;
     Window = SDL_CreateWindow(NULL, 0, 0, windowWidth, windowHeight, SDL_WINDOW_FULLSCREEN);
-    if (!Window) {
+    if (!Window) 
+    {
         std::cerr << "Error creating SDL window" << std::endl;
         return false;
     }
     Renderer = SDL_CreateRenderer(Window, -1, SDL_RENDERER_ACCELERATED | SDL_RENDERER_PRESENTVSYNC);
-    if (!Renderer) {
+    if (!Renderer) 
+    {
         std::cerr << "Error creating SDL renderer" << std::endl;
         return false;
     }
     return true;
 }
 
-void Graphics::ClearScreen(Uint32 color) {
+void Graphics::ClearScreen(Uint32 color) 
+{
     SDL_SetRenderDrawColor(Renderer, color >> 16, color >> 8, color, 255);
     SDL_RenderClear(Renderer);
 }
 
-void Graphics::RenderFrame() {
+void Graphics::RenderFrame() 
+{
     SDL_RenderPresent(Renderer);
 }
 
-void Graphics::DrawLine(int x0, int y0, int x1, int y1, Uint32 color) {
+void Graphics::DrawLine(int x0, int y0, int x1, int y1, Uint32 color) 
+{
     lineColor(Renderer, x0, y0, x1, y1, color);
 }
 
-void Graphics::DrawCircle(int x, int y, int radius, float angle, Uint32 color) {
+void Graphics::DrawCircle(int x, int y, int radius, float angle, Uint32 color) 
+{
     circleColor(Renderer, x, y, radius, color);
     lineColor(Renderer, x, y, x + cos(angle) * radius, y + sin(angle) * radius, color);
 }
 
-void Graphics::DrawFillCircle(int x, int y, int radius, Uint32 color) {
+void Graphics::DrawFillCircle(int x, int y, int radius, Uint32 color) 
+{
     filledCircleColor(Renderer, x, y, radius, color);
 }
 
-void Graphics::DrawRect(int x, int y, int width, int height, Uint32 color) {
+void Graphics::DrawRect(int x, int y, int width, int height, Uint32 color) 
+{
     lineColor(Renderer, x - width / 2.0, y - height / 2.0, x + width / 2.0, y - height / 2.0, color);
     lineColor(Renderer, x + width / 2.0, y - height / 2.0, x + width / 2.0, y + height / 2.0, color);
     lineColor(Renderer, x + width / 2.0, y + height / 2.0, x - width / 2.0, y + height / 2.0, color);
@@ -69,8 +78,9 @@ void Graphics::DrawFillRect(int x, int y, int width, int height, Uint32 color) {
     boxColor(Renderer, x - width / 2.0, y - height / 2.0, x + width / 2.0, y + height / 2.0, color);
 }
 
-void Graphics::DrawPolygon(int x, int y, const std::vector<Vector2D>& vertices, Uint32 color) {
-    for (int i = 0; i < vertices.size(); i++) {
+void Graphics::DrawPolygon(int x, int y, const std::vector<Vector2D>& vertices, Uint32 color) 
+{
+    for (int i = 0; i < static_cast<int>(vertices.size()); i++) {
         int currIndex = i;
         int nextIndex = (i + 1) % vertices.size();
         lineColor(Renderer, vertices[currIndex].x, vertices[currIndex].y, vertices[nextIndex].x, vertices[nextIndex].y, color);
@@ -78,26 +88,32 @@ void Graphics::DrawPolygon(int x, int y, const std::vector<Vector2D>& vertices, 
     filledCircleColor(Renderer, x, y, 1, color);
 }
 
-void Graphics::DrawFillPolygon(int x, int y, const std::vector<Vector2D>& vertices, Uint32 color) {
+void Graphics::DrawFillPolygon(int x, int y, const std::vector<Vector2D>& vertices, Uint32 color) 
+{
     std::vector<short> vx;
     std::vector<short> vy;
-    for (int i = 0; i < vertices.size(); i++) {
+
+    for (int i = 0; i < static_cast<int>(vertices.size()); i++) 
+    {
         vx.push_back(static_cast<int>(vertices[i].x));
     }
-    for (int i = 0; i < vertices.size(); i++) {
+    for (int i = 0; i < static_cast<int>(vertices.size()); i++) 
+    {
         vy.push_back(static_cast<int>(vertices[i].y));
     }
     filledPolygonColor(Renderer, &vx[0], &vy[0], vertices.size(), color);
     filledCircleColor(Renderer, x, y, 1, 0xFF000000);
 }
 
-void Graphics::DrawTexture(int x, int y, int width, int height, float rotation, SDL_Texture* texture) {
+void Graphics::DrawTexture(int x, int y, int width, int height, float rotation, SDL_Texture* texture) 
+{
     SDL_Rect dstRect = {x - (width / 2), y - (height / 2), width, height};
     float rotationDeg = rotation * 57.2958;
     SDL_RenderCopyEx(Renderer, texture, NULL, &dstRect, rotationDeg, NULL, SDL_FLIP_NONE);
 }
 
-void Graphics::CloseWindow(void) {
+void Graphics::CloseWindow(void) 
+{
     SDL_DestroyRenderer(Renderer);
     SDL_DestroyWindow(Window);
     SDL_Quit();
